requirements.txt
discord.py-self>=2.1.0,<3.0.0
python-dotenv>=1.0,<2.0
